% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgmin-package.R
\docType{package}
\name{pkgmin-package}
\alias{pkgmin}
\alias{pkgmin-package}
\title{pkgmin: Structure d'un package simple sous R}
\description{
Le package \code{pkgmin} illustre la structure d'un package...
}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr}

}
\keyword{internal}
