% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy.R
\docType{data}
\name{xy}
\alias{xy}
\title{Exemple d'un jeu de donnees interne}
\format{Un data frame avec 9 lignes et 2 colonnes :
\describe{
\item{x}{valeurs tirees aleatoirement dans une distribution normale}
\item{y}{sequences de 2 a 9 avec une valeur manquante}
}}
\source{
Random Number Generator (RNG): 1637
}
\usage{
xy
}
\description{
Un jeu de donnees fictif illustrant la construction d'un
exemple reproductible interne
}
\keyword{datasets}
